% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{interactive_multi_3d}
\alias{interactive_multi_3d}
\title{Interactive Multi-Drop 3D Zone / Interaktive 3D-Mehrfach-Klick-Zone}
\usage{
interactive_multi_3d()
}
\value{
\strong{DEUTSCH:} Keinen Rückgabewert, zeichnet eine interaktive 3D-Perspektive nach 3 Mausklicks. / \strong{ENGLISH:} No return value, plots a 3D perspective after 3 interactive mouse clicks.
}
\description{
Allows the user to click 3 times to create overlapping waves in the 3D landscape.
Erlaubt 3 Mausklicks, um sich ueberlagernde Wellen in einer 3D-Landschaft zu erzeugen.
}
