% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{interactive_octawave}
\alias{interactive_octawave}
\title{Interactive Drop Zone / Interaktives Mausklick-Cockpit}
\usage{
interactive_octawave()
}
\value{
\strong{DEUTSCH:} Keinen Rückgabewert, wird für interaktive Plots im Grafikfenster genutzt. / \strong{ENGLISH:} No return value, called for interactive plotting side effects.
}
\description{
Allows the user to click into the plot to trigger a new wave at that exact coordinate.
Erlaubt es, per Mausklick eine neue Welle an der gewaehlten Koordinate auszuloesen.
}
