% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{octa_vertices}
\alias{octa_vertices}
\title{Geometrische Grundstruktur des Quanten-Oktaeders / Geometric Core Structure of the Quantum Octahedron}
\usage{
octa_vertices(radius = 1)
}
\arguments{
\item{radius}{Der Abstand der Ecken vom Nullpunkt (Skalierung des Quanten-Kerns) / The distance of the vertices from the origin (scaling of the quantum core).}
}
\value{
\strong{DEUTSCH:} Eine Matrix mit 6 Zeilen und 3 Spalten (x, y, z), die die räumlichen Koordinaten der Oktaeder-Ecken enthält. / \strong{ENGLISH:} A matrix with 6 rows and 3 columns (x, y, z) containing the spatial coordinates of the octahedral vertices.
}
\description{
\strong{DEUTSCH:} Berechnet die exakten 3D-Koordinaten der 6 Oktaeder-Scheitelpunkte im Raum.

\strong{ENGLISH:} Computes the exact 3D coordinates of the 6 octahedral vertices in space.
}
