% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_sync3d.R
\name{prep_sync3d}
\alias{prep_sync3d}
\title{Prepare Waves for Katharina's sync3d Engine}
\usage{
prep_sync3d(analysis_results)
}
\arguments{
\item{analysis_results}{The output list generated by the wave2lyze function.}
}
\description{
Generates the node data frame and the topological edge matrix for add_synchronized_3d_edges.
}
