% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mri_simulation.R
\name{run_mri_simulation}
\alias{run_mri_simulation}
\title{Run Dynamic Quantum MRI Scans}
\usage{
run_mri_simulation(sleep_time = 0.05)
}
\arguments{
\item{sleep_time}{Numeric. The pause duration between sequential slices in seconds. Defaults to 0.05.}
}
\description{
Launches the automated 80-step Fullerene and Octawave animation sequence directly in the R session.
}
