% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{export_all_mri_scans}
\alias{export_all_mri_scans}
\title{Exportiert eine komplette Serie von Fullerene- und Octawave-MRT-Schichten (PNGs + PDFs)}
\usage{
export_all_mri_scans(start_z = -2, end_z = 2, schritte = 10)
}
\arguments{
\item{start_z}{Die Start-Hoehe (z.B. -2)}

\item{end_z}{Die End-Hoehe (z.B. 2)}

\item{schritte}{Wie viele Bilder pro Serie gezeichnet werden sollen}
}
\description{
Exportiert eine komplette Serie von Fullerene- und Octawave-MRT-Schichten (PNGs + PDFs)
}
