% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{export_octawave_series}
\alias{export_octawave_series}
\title{Export 3D Octawave Snapshot Series / 3D-Bilder-Serie exportieren}
\usage{
export_octawave_series(count = 5)
}
\arguments{
\item{count}{Number of images to export / Anzahl der Bilder, die gespeichert werden sollen.}
}
\description{
Saves a series of 3D snapshots as PNG files to visualize the wave progression.
Speichert eine Serie von 3D-Momentaufnahmen als PNG-Dateien, um das Fortschreiten zu zeigen.
}
