% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{octa_vertices}
\alias{octa_vertices}
\title{Geometrische Grundstruktur des Quanten-Oktaeders}
\usage{
octa_vertices(radius = 1)
}
\arguments{
\item{radius}{Der Abstand der Ecken vom Nullpunkt (Skalierung des Quanten-Kerns).}
}
\description{
Berechnet die exakten 3D-Koordinaten der 6 Oktaeder-Scheitelpunkte im Raum.
}
