% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{octa_vertices}
\alias{octa_vertices}
\title{Geometrische Grundstruktur des Quanten-Oktaeders / Geometric Core Structure of the Quantum Octahedron}
\usage{
octa_vertices(radius = 1)
}
\arguments{
\item{radius}{Der Abstand der Ecken vom Nullpunkt (Skalierung des Quanten-Kerns) / The distance of the vertices from the origin (scaling of the quantum core).}
}
\description{
\link{DEUTSCH} Berechnet die exakten 3D-Koordinaten der 6 Oktaeder-Scheitelpunkte im Raum.

\link{ENGLISH} Computes the exact 3D coordinates of the 6 octahedral vertices in space.
}
