% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{octa_wave_field}
\alias{octa_wave_field}
\title{neuartige Oktaeder-Quantenwelle mit Raumdämpfung}
\usage{
octa_wave_field(x, y, z, freq = 2, radius = 1.5, damping = 0.2)
}
\arguments{
\item{x, y, z}{Die Koordinaten des Punktes im Raum.}

\item{freq}{Die Schwingungsfrequenz der Wellenkomponenten.}

\item{radius}{Der Radius des zugrundeliegenden Oktaeders.}

\item{damping}{Der Dämpfungsfaktor (wie schnell die Welle im Raum abklingt).}
}
\description{
Berechnet die Superposition von Wellen, die von den 6 Ecken eines Oktaeders ausgestrahlt werden, inklusive einer quantenmechanischen Amplitudendämpfung.
}
