% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_octawave}
\alias{plot_octawave}
\title{Visualisierung der octawave Quantenwelle}
\usage{
plot_octawave(resolution = 200, range = 8)
}
\arguments{
\item{resolution}{Die Auflösung des Gitters (höhere Werte bedeuten feinere Bilder).}

\item{range}{Der sichtbare Bereich auf der X- und Y-Achse.}
}
\description{
Erstellt einen automatischen 2D-Schnitt durch das Oktaeder-Wellenfeld im klassischen MATLAB-Stil.
}
