% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_octawave_3d}
\alias{plot_octawave_3d}
\title{Interaktive 3D-Visualisierung der Oktaeder-Welle / Interactive 3D Visualization of the Octahedral Wave}
\usage{
plot_octawave_3d(resolution = 35, range = 5)
}
\arguments{
\item{resolution}{Die Feinheit des 3D-Gitters (z.B. 30 oder 40) / The density of the 3D grid (e.g., 30 or 40).}

\item{range}{Der sichtbare Raumbereich (X, Y und Z) / The visible spatial range (X, Y, and Z).}
}
\description{
\strong{DEUTSCH]:} Erstellt ein dreidimensionales, interaktives Volumenmodell des Quantenfeldes.

\strong{ENGLISH:} Generates a three-dimensional, interactive volume model of the quantum field.
}
